refactor: changed how the former validateWordleGuess and associated methods work
validateWordleGuess has been updated to return void, since it now throws an exception if the entered guess is too short. Otherwise, a new method will be called that in return calls two methods with one checking if the guess is equal to the solution and if not, the other will update each WordlePanel according to the game rules.